Ditch translucency as a configuration option.
This is slightly complicated because Boom WADs can use translucency on
linedefs, and BEX patches can apply the translucency bit to objects.
So remove the translucency bit that was retrofitted to Doom objects and
delete the options from the menu (they're still in the config file).

Rationale: this is a feature that seemed like a cool bit of glitz to add
to Doom back in the 90s but I don't think it's aged well. It doesn't
really add anything to the game, and actually has a negative effect on
visuals since objects fireballs are dimmer and less visible.