Remove keyed-door map colors.
These are perhaps one of the most worthwhile extensions added by
Boom, as it's actually useful to be able to re-find the colored door
you saw earlier when you pick up the key.

However, I don't think the case is strong enough for keeping them;
red and yellow conflict with the original Doom map colors, and if
necessary it's still possible to use a map marker to remember where
the doors are for later.