Get rid of the Boom HUD.
A potentially controversial change, since we don't have a replacement
HUD yet, and a fullscreen HUD is definitely worth having.

The Boom HUD really is just not good. I've heard so many people
express their dislike for it, that I think it's better to not have a
HUD than to keep this one. At least removing it adds an incentive to
develop a replacement.

This first step just removes the easiest bits of the HUD. Further
changes will dig deeper.