Get rid of key_backspace/key_enter.
Is there really anyone who wants to rebind their multiplayer chat
keys so they press something other than backspace to delete characters,
or something other than enter to send the message? I'm going to go out
on a limb and say probably not.

I believe this also fixes a bug with the heads-up code: key_enter is
used in place of KEYD_ENTER where \n is actually used as part of the
chat message protocol.