Swap "game" and "menus" sections on key bindings screen.
This is the first step for rearranging the key bindings into a better
order. On balance, the "game" shortcuts, with quickload/save bindings,
are something more useful to show on the first screen. But leave the
"quit game" binding behind since it's less likely to be changed. This
also frees up some space on the first screen, which will be useful.