Replace realtic_clock_rate with command line arg.
Having discussed this with some speedrunners I've had the feedback
that having to configure this as a config file option is actually very
frustrating. Usually this is something that is just wanted temporarily
for one run of the program, and it's easy to forget to change it back
to 100% once it's no longer needed, which can result in unintentionally
recording TAS demos. With that in mind, it's a more natural fit for
this to be a command line argument, and we can get rid of the option
from the menu and config file.