Only show double key indicator on levels that need it.
99+% of levels only use the normal keyed door types where there is
no distinction between keycards and skull keys. So on level load, check
all the special linedef types to see if any of them is a generalized
type that distinguishes between door types. If not, don't show the double
key indicator on the status bar.

MBF added the sts_traditional_keys option to hide the double key display,
probably because the icon itself is kind of cramped and ugly (it's hard to
convey that much information in such a tiny icon). I think this is a better
option.