Get rid of monster_backing option.
"Monster backing out" makes monsters back away when confronted by an
enemy with a melee attack. There's nothing in log_lee.txt about why
this was added but I'm guessing it was to even the score a bit for
monsters when confronted by vicious dogs on the attack. It's pretty
cute to see Cyberdemons back away from a pack of dogs. Anyway, the
DOGS code was already removed so this seems a bit superfluous. Keep
for demo compatibility but delete the option.