CMake build system
This should help with building the port in modern versions of Visual
Studio.  Mostly taken from Chocolate Doom, and there is probably more
work that could be done to it, but it at least builds.

This commit also contains a few bugfixes that prevented the project from
being built, nothing major.