Get rid of "Setup" menu.
It now only contains two items: Key Bindings and Chat Strings. With
that in mind, we can delete the menu and Key Bindings can take its
spot on the Options menu.

This does temporarily leave the Chat Strings menu orphaned, but I
plan to bring the link to it back as part of a rework of the main
options menu.