Partially revert "Delete all DOGS code."
This reverts commit 789b5d3cff54275c18bd26297a3460ea19309072.

Removing the MBF dogs seemed like a no-brainer. However, various mods
over the years (such as Eviternity) have taken advantage of the extra
frames that were added for the dogs, so we have to retain them.

This is a partial revert of the previous commit - the option to
configure helper dogs is still gone, but all the logic is added back.